requirements.txt
streamlit
pandas
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
